requirements.txt
slack-bolt==1.20.0
requests
python-dotenv   # optional, for loading .env in local dev
